renaming AppContext.js to .jsx, removing file extension from imports also
